+++ src/index.ts
@@ -4,6 +4,7 @@ export * from "./deprecationRegistry.js";
 export * from "./validateEnvelope.js";
 export * from "./envelope.js";
 export * from "./flagSemantics.js";
+export * from "./fieldExtraction.js";
 export * from "./conformance.js";
 export * from "./conformanceProfiles.js";
 export * from "./compliance.js";